set and clean updatePlaylist funct w/ cloudinary

+++ controllers/playlist.js
@@ -1,6 +1,9 @@
 const Playlist = require("../models/Playlist");
 const User = require("../models/User");
-const { uploadImage } = require("../utils/cloudinary");
+const {
+  uploadImage,
+  deleteImage
+} = require("../utils/cloudinary");
 const fs = require("fs-extra");
 
 const getPlaylists = async (req, res) => {
@@ -23,7 +26,11 @@ const getPlaylists = async (req, res) => {
 };
 
 const followPlaylists = async (req, res) => {
-  const { loggedUserId, playlistId, isAdded } = req.body;
+  const {
+    loggedUserId,
+    playlistId,
+    isAdded
+  } = req.body;
 
   try {
     const loggedUser = await User.findOne({
@@ -38,18 +45,15 @@ const followPlaylists = async (req, res) => {
         },
       });
 
-      await Playlist.findOneAndUpdate(
-        {
-          _id: playlistId,
-        },
-        {
-          $addToSet: {
-            followedBy: {
-              $each: [loggedUserId],
-            },
+      await Playlist.findOneAndUpdate({
+        _id: playlistId,
+      }, {
+        $addToSet: {
+          followedBy: {
+            $each: [loggedUserId],
           },
-        }
-      );
+        },
+      });
 
       return res.status(200).json({
         ok: true,
@@ -66,18 +70,15 @@ const followPlaylists = async (req, res) => {
         },
       });
 
-      await Playlist.findOneAndUpdate(
-        {
-          _id: playlistId,
-        },
-        {
-          $pull: {
-            followedBy: {
-              $in: loggedUserId,
-            },
+      await Playlist.findOneAndUpdate({
+        _id: playlistId,
+      }, {
+        $pull: {
+          followedBy: {
+            $in: loggedUserId,
           },
-        }
-      );
+        },
+      });
       return res.status(200).json({
         ok: true,
         loggedUserId,
@@ -95,7 +96,10 @@ const followPlaylists = async (req, res) => {
 };
 
 const createPlaylist = async (req, res) => {
-  const { name, isPrivate } = JSON.parse(req.body.imagePlaylistData);
+  const {
+    name,
+    isPrivate
+  } = JSON.parse(req.body.imagePlaylistData);
 
   const userId = req.params.userId;
   const file = req.files[0];
@@ -143,7 +147,10 @@ const createPlaylist = async (req, res) => {
 };
 
 const deletePlaylist = async (req, res) => {
-  const { loggedUserId, playlistId } = req.body;
+  const {
+    loggedUserId,
+    playlistId
+  } = req.body;
 
   try {
     const loggedUser = await User.findOne({
@@ -163,22 +170,28 @@ const deletePlaylist = async (req, res) => {
         playlists: playlistId,
       },
     });
+
+    const response = await deleteImage(playlistToDelete.thumbnailCloudinaryId)
+    if (!response.result === "ok") {
+      return res.status(503).json({
+        ok: false,
+        msg: response.result
+      });
+    }
+
     await Playlist.findByIdAndDelete(playlistId).then((deletedPlaylist) => {
       return res.status(200).json({
         ok: true,
         deletedPlaylist,
       });
     });
-    await User.updateMany(
-      {
+    await User.updateMany({
+      followedPlaylists: playlistId,
+    }, {
+      $pull: {
         followedPlaylists: playlistId,
       },
-      {
-        $pull: {
-          followedPlaylists: playlistId,
-        },
-      }
-    );
+    });
   } catch (error) {
     console.log(error);
     return res.status(503).json({
@@ -188,41 +201,10 @@ const deletePlaylist = async (req, res) => {
   }
 };
 
-const updatePlaylist = async (req, res) => {
-  const { loggedUserId, playlistId, newName, thumbnailUrl } = req.body;
-  try {
-    const playlistToUpdate = await Playlist.findOne({
-      _id: playlistId,
-    });
-    if (playlistToUpdate.user.toString() !== loggedUserId) {
-      return res.status(401).json({
-        ok: false,
-        message: "You are not the owner of this playlist",
-      });
-    }
-    const oldName = playlistToUpdate.name;
-    await playlistToUpdate.updateOne({
-      name: newName,
-    });
-    await playlistToUpdate.updateOne({
-      thumbnail: thumbnailUrl,
-    });
-    return res.status(200).json({
-      ok: true,
-      playlistToUpdate,
-      oldName,
-      newName,
-    });
-  } catch (error) {
-    return res.status(503).json({
-      ok: false,
-      msg: "Oops, something happened",
-    });
-  }
-};
-
 const getPlaylistById = async (req, res) => {
-  const { id } = req.params;
+  const {
+    id
+  } = req.params;
 
   if (id.length !== 24) {
     return res.status(200).json({
@@ -233,7 +215,12 @@ const getPlaylistById = async (req, res) => {
   try {
     const playlist = await Playlist.findOne({
       _id: id,
-    }).populate({ path: "tracks", populate: { path: "artist" } });
+    }).populate({
+      path: "tracks",
+      populate: {
+        path: "artist"
+      }
+    });
 
     if (!playlist) {
       return res.status(404).json({
@@ -256,7 +243,11 @@ const getPlaylistById = async (req, res) => {
 };
 
 const isPrivate = async (req, res) => {
-  const { loggedUserId, playlistId, isPrivate } = req.body;
+  const {
+    loggedUserId,
+    playlistId,
+    isPrivate
+  } = req.body;
 
   try {
     const playlistToUpdate = await Playlist.findOne({
@@ -274,16 +265,13 @@ const isPrivate = async (req, res) => {
     });
 
     if (!isPrivate) {
-      await User.updateMany(
-        {
+      await User.updateMany({
+        followedPlaylists: playlistId,
+      }, {
+        $pull: {
           followedPlaylists: playlistId,
         },
-        {
-          $pull: {
-            followedPlaylists: playlistId,
-          },
-        }
-      );
+      });
     }
 
     return res.status(200).json({
@@ -298,7 +286,12 @@ const isPrivate = async (req, res) => {
   }
 };
 const addTracks = async (req, res) => {
-  const { loggedUserId, playlistId, trackId, isAdded } = req.body;
+  const {
+    loggedUserId,
+    playlistId,
+    trackId,
+    isAdded
+  } = req.body;
   try {
     const playlistToUpdate = await Playlist.findOne({
       _id: playlistId,
@@ -336,7 +329,10 @@ const addTracks = async (req, res) => {
 };
 
 const duplicatePlaylist = async (req, res) => {
-  const { loggedUserId, playlistId } = req.body;
+  const {
+    loggedUserId,
+    playlistId
+  } = req.body;
   try {
     const playlistToDuplicate = await Playlist.findOne({
       _id: playlistId,
@@ -368,7 +364,13 @@ const duplicatePlaylist = async (req, res) => {
       );
 
       if (missingTracks.length !== 0) {
-        playlist.updateOne({ $push: { tracks: { $each: missingTracks } } });
+        playlist.updateOne({
+          $push: {
+            tracks: {
+              $each: missingTracks
+            }
+          }
+        });
         return res.status(200).json({
           ok: true,
           msg: "This playlist is already in your library. The missing songs have been added to the corresponding playlist.",
@@ -392,7 +394,11 @@ const duplicatePlaylist = async (req, res) => {
       copyFrom: playlistId,
     });
     await newPlaylist.save();
-    await loggedUser.updateOne({ $push: { playlists: newPlaylist._id } });
+    await loggedUser.updateOne({
+      $push: {
+        playlists: newPlaylist._id
+      }
+    });
     return res.status(201).json({
       ok: true,
       msg: `The playlist ${playlistToDuplicate.name} have been duplicated!`,
@@ -408,25 +414,63 @@ const duplicatePlaylist = async (req, res) => {
   }
 };
 
-const updatePlaylistName = async (req, res) => {
-  const { playlistId } = req.params;
-  const { namePlaylist } = req.body;
+const updatePlaylist = async (req, res) => {
+  const {
+    name
+  } = JSON.parse(req.body.imagePlaylistData);
+  const playlistId = req.params.playlistId
+  const file = req.files[0]
 
   try {
-    const playlist = await Playlist.findById(playlistId);
-
-    if (!playlist) {
-      return res.status(404).json({ message: "Playlist not found" });
+    if (!file) {
+      return res.status(503).json({
+        ok: false,
+        msg: "No files uploaded",
+      });
     }
-    playlist.name = namePlaylist;
-    await playlist.save();
+    if (file) {
+      //Upload thumbnail to Cloudinary
+      const resultImage = await uploadImage(file.path)
+      const url = resultImage.secure_url
+      const cloudinaryId = resultImage.public_id
 
-    res.status(200).json(playlist);
+      const playlistBeforeUpdate = await Playlist.findOneAndUpdate({
+        _id: playlistId
+      }, {
+        $set: {
+          name: name,
+          thumbnail: url,
+          thumbnailCloudinaryId: cloudinaryId
+        },
+      }, {
+        new: false
+      })
+
+      const response = await deleteImage(playlistBeforeUpdate.thumbnailCloudinaryId)
+      if (!response.result === "ok") {
+        return res.status(503).json({
+          ok: false,
+          msg: response.result
+        });
+      }
+
+      await fs.unlink(file.path)
+
+      return res.status(201).json({
+        ok: true,
+        newName: name,
+        thumbnail: url,
+      });
+    }
   } catch (error) {
-    console.error(error);
-    res.status(500).json({ message: "Internal server error" });
+    console.log(error);
+    return res.status(503).json({
+      ok: false,
+      msg: "Oops, something happened",
+    });
   }
 };
+
 
 module.exports = {
   getPlaylists,
@@ -438,5 +482,4 @@ module.exports = {
   isPrivate,
   duplicatePlaylist,
   addTracks,
-  updatePlaylistName,
 };
